requirements.txt
# Core dependencies
openai>=1.109.1
playwright>=1.55.0
python-dotenv>=1.1.1
requests>=2.32.5
flask>=2.3.0

# Weaviate and vector database
weaviate-client>=4.17.0

# HTTP and networking
httpx>=0.28.1
httpcore>=1.0.9
websockets>=10.4

# Data validation and parsing
pydantic>=2.11.9
pydantic-core>=2.33.2

# Async and concurrency
anyio>=4.11.0
greenlet>=3.2.4

# Security and authentication
cryptography>=46.0.1
certifi>=2025.8.3
Authlib>=1.6.4

# Utilities
colorama>=0.4.6
tqdm>=4.67.1
validators>=0.35.0
packaging>=25.0

# Type hints and annotations
typing-extensions>=4.15.0
annotated-types>=0.7.0

# Development and debugging
deprecation>=2.1.0
exceptiongroup>=1.3.0

# System utilities
distro>=1.9.0
appdirs>=1.4.4
importlib-metadata>=8.7.0
zipp>=3.23.0

# Text processing
charset-normalizer>=3.4.3
idna>=3.10

# Protocol buffers and gRPC
protobuf>=6.32.1
grpcio>=1.75.1

# HTTP/2 support
h11>=0.16.0
sniffio>=1.3.1

# JSON processing
jiter>=0.11.0

# Event handling
pyee>=13.0.0

# C extensions
cffi>=2.0.0
pycparser>=2.23.0

# URL handling
urllib3>=2.5.0

# Type inspection
typing-inspection>=0.4.1